pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "envradar"
version = "0.1.0"
description = "Detect undocumented, unused, and drifting environment variables in repositories."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["dotenv", "env", "cli", "developer-tools", "github", "open-source"]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[project.scripts]
envradar = "envradar.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
